Do not lose old state
Previously, the state had to be rebuild by all providers on each run.
While this was nice from a "set the correct state" point of view, it
caused some issues with long term state / cross provider state, so it
makes more sense to prime the state with the old values.