Support overriding Ingress Service
Currently, ingress-nginx does not copy hostnames into Ingress Status
blocks if there is an IP address available. This is problematic for
cloudfront, as it does not support IP address origins.

This change adds a small shim which can be applied when running the
controller. When passing --ingress-service, the Ingress Controller will
ignore the status block on any Ingress objects, and will instead read
the status directly from the given Service in the given namespace.

A [pull request][1] is open with ingress-nginx to fix this issue
upstream.

[1]: https://github.com/kubernetes/ingress-nginx/pull/7464